requirements.txt
flask
pymongo
sentence-transformers
torch
python-dotenv
